added a naming for self-cal test
